updated make_network
